# openapi/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "message": {
            "type": "string"
          },
          "reason": {
            "description": "Machine-readable classifier: NotFound, BadRequest, Conflict, ManagedByGitOps, Internal, ServiceUnavailable.",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Thin REST/WS API the GUI consumes. Reads flow through a controller-runtime shared informer cache; writes go directly to the k8s API server (§CLAUDE.md §7). §7 managed-by policy is enforced here for PATCH/DELETE on Tests; TestRun creation is always allowed regardless of the referenced Test's ownership.",
    "title": "kubetest-alt API",
    "version": "v1alpha1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs/{id}/artifacts": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "contentType": {
                        "type": "string"
                      },
                      "path": {
                        "type": "string"
                      },
                      "sizeBytes": {
                        "type": "integer"
                      }
                    },
                    "required": [
                      "path"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List the run's artifacts (recorded refs; falls back to listing the run's artifacts/ prefix when none were recorded)."
      },
      "parameters": [
        {
          "description": "TestRun name or UID.",
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "Target namespace. Required for single-object requests when the server runs cluster-wide; optional (and must match) when it is scoped. On list endpoints, omitting it on a cluster-wide server lists all namespaces.",
          "in": "query",
          "name": "namespace",
          "required": false,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/runs/{id}/artifacts/{path}": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "expiresIn": {
                      "type": "integer"
                    },
                    "url": {
                      "format": "uri",
                      "type": "string"
                    }
                  },
                  "required": [
                    "url",
                    "expiresIn"
                  ],
                  "type": "object"
                }
              },
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Artifact bytes, or {url, expiresIn} with ?presign=1."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "OK"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Stream an artifact's bytes (default; served with nosniff + CSP sandbox, ?download=1 for attachment), or with ?presign=1 return a presigned object-store URL. Path traversal (../, absolute paths, backslashes) rejected 400."
      },
      "parameters": [
        {
          "description": "TestRun name or UID.",
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "Relative artifact path (e.g. results/junit.xml).",
          "in": "path",
          "name": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "description": "Target namespace. Required for single-object requests when the server runs cluster-wide; optional (and must match) when it is scoped. On list endpoints, omitting it on a cluster-wide server lists all namespaces.",
          "in": "query",
          "name": "namespace",
          "required": false,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  }
}